Move main api file to slightly more sensible location
